requirements.txt
celery==5.2.7
Django==3.2
django-cors-headers==3.4.0
djangorestframework==3.11.0
mysqlclient==2.1.0
djangorestframework-simplejwt==4.6.0
gunicorn==19.9.0
Pillow==9.2.0
slackclient==2.5.0
pid==3.0.0
xlsxwriter==1.2.8
pdfplumber==0.5.23
pandas==1.4.2
numpy==1.22.2
PyJWT==1.7.1
django-filter==2.4.0
django-auto-prefetching==0.1.8
xlrd==1.2.0
openpyxl==3.0.7
reportlab==3.5.67
requests==2.26.0
holidays==0.12
drf-yasg==1.20.0
whitenoise==6.2.0